changed packet structure and added console script
