+++ Dockerfile
@@ -80,6 +80,9 @@ COPY vid.mp4 /home/video_cap
 RUN python3.10 -m pip install --upgrade pip build twine && \
   python3.10 -m pip install 'pkgconfig>=1.5.1' 'numpy>=1.17.0'
 
-RUN python3.10 setup.py install
+RUN python3.10 -m pip install .
+
+# that is where the "extract_mv" script is located
+ENV PATH="$PATH:/opt/_internal/cpython-3.10.2/bin"
 
 CMD ["sh", "-c", "tail -f /dev/null"]
