+++ setup.py
@@ -1,4 +1,4 @@
-from setuptools import setup, Extension
+from setuptools import find_packages, setup, Extension
 import pkgconfig
 import numpy as np
 
@@ -6,7 +6,7 @@ d = pkgconfig.parse('libavformat libswscale opencv4')
 
 print("Numpy dir: ", np.get_include())
 
-mv_extractor = Extension('mv_extractor',
+mv_extractor = Extension('mv_extractor.videocap',
        include_dirs = [
               *d['include_dirs'],
               np.get_include()
@@ -14,10 +14,10 @@ mv_extractor = Extension('mv_extractor',
        library_dirs = d['library_dirs'],
        libraries = d['libraries'],
        sources = [
-              'src/py_video_cap.cpp',
-              'src/video_cap.cpp',
-              'src/time_cvt.cpp',
-              'src/mat_to_ndarray.cpp'
+              'src/mv_extractor/py_video_cap.cpp',
+              'src/mv_extractor/video_cap.cpp',
+              'src/mv_extractor/time_cvt.cpp',
+              'src/mv_extractor/mat_to_ndarray.cpp'
        ],
        extra_compile_args = ['-std=c++11'],
        extra_link_args = ['-fPIC', '-Wl,-Bsymbolic'])
@@ -31,7 +31,19 @@ setup(name='mv_extractor',
        description=('Reads video frames and MPEG-4/H.264 motion vectors.'),
        keywords=['motion vector', 'video capture', 'mpeg4', 'h.264', 'compressed domain'],
        ext_modules=[mv_extractor],
-       packages=['mv_extractor'],
-       package_dir={'mv_extractor': 'src'},
+       packages=find_packages(where='src'),
+       package_dir={'': 'src'},
+       #package_dir={'mv_extractor': 'src'},
+       #package_data={
+       #       'mv_extractor': [
+       #              'vid.mp4',
+       #              ''
+       #       ],
+       #},
+       entry_points={
+              'console_scripts': [
+                     'extract_mvs=mv_extractor.__main__:main',
+              ],
+       },
        python_requires='>=3.8, <4',
        install_requires=['pkgconfig>=1.5.1', 'numpy>=1.17.0', 'opencv-python>=4.1.0.25,<4.6'])
